% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blocos}
\alias{blocos}
\title{Blocos de países}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 323 rows and 5 columns.
}
\usage{
blocos
}
\description{
Blocos de países
}
\keyword{data}
