% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download}
\alias{download}
\alias{baixar_comex}
\alias{baixar_importacao}
\alias{baixar_exportacao}
\title{Download de dados do comércio exterior}
\usage{
baixar_comex(ano, tipo, destino)

baixar_importacao(ano, destino)

baixar_exportacao(ano, destino)
}
\arguments{
\item{ano}{ano}

\item{tipo}{exportação (exp) ou importação (imp)}

\item{destino}{}
}
\value{
TRUE/FALSE sucesso da requisição
}
\description{
Download de dados do comércio exterior
}
