% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{ler_comex}
\alias{ler_comex}
\title{Ler dados comércio exterior}
\usage{
ler_comex(caminho)
}
\arguments{
\item{caminho}{arquivo de import/export}
}
\value{
um data.frame
}
\description{
Ler dados comércio exterior
}
