% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ncm}
\alias{ncm}
\title{NCM (Nomenclatura Comum do Mercosul)}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 13111 rows and 14 columns.
}
\usage{
ncm
}
\description{
NCM (Nomenclatura Comum do Mercosul)
}
\keyword{data}
