% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{paises}
\alias{paises}
\title{Países}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 281 rows and 6 columns.
}
\usage{
paises
}
\description{
Países
}
\keyword{data}
