% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sh}
\alias{sh}
\title{SH (Sistema Harmonizado)}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 6308 rows and 16 columns.
}
\usage{
sh
}
\description{
SH (Sistema Harmonizado)
}
\keyword{data}
